Auto layout helpers!

+++ KTCategories/UIColor+KTExtras.h
@@ -11,7 +11,20 @@
 
 @interface UIColor (KTExtras)
 
-+ (UIColor *)greyColorWithBrightness:(CGFloat)brightness;
-+ (UIColor *)greyColorWithBrightness:(CGFloat)brightness alpha:(float)alpha;
+/**
+ * Generate a shade of grey from HSB(0,0,brightness,1)
+ *
+ * @param brightness The brightness value for an HSB color (e.g 0.5f)
+ */
++ (UIColor *)greyColorWithBrightness:(float)brightness;
+
+/**
+ * Generate a shade of grey with alpha from HSB(0,0,brightness,alpha)
+ *
+ * @param brightness The brightness value for an HSB color (e.g 0.5f)
+ * @param alpha The alpha value for an HSB color (e.g 0.5f)
+ */
++ (UIColor *)greyColorWithBrightness:(float)brightness
+                               alpha:(float)alpha;
 
 @end
