+++ KTCategories/UIColor+KTExtras.m
@@ -10,14 +10,21 @@
 
 @implementation UIColor (KTExtras)
 
-+ (UIColor *)greyColorWithBrightness:(CGFloat)brightness
++ (UIColor *)greyColorWithBrightness:(float)brightness
 {
-  return [UIColor colorWithHue:0 saturation:0 brightness:brightness alpha:1.f];
+  return [UIColor colorWithHue:0
+                    saturation:0
+                    brightness:brightness
+                         alpha:1.f];
 }
 
-+ (UIColor *)greyColorWithBrightness:(CGFloat)brightness alpha:(float)alpha
++ (UIColor *)greyColorWithBrightness:(float)brightness
+                               alpha:(float)alpha
 {
-  return [UIColor colorWithHue:0 saturation:0 brightness:brightness alpha:alpha];
+  return [UIColor colorWithHue:0
+                    saturation:0
+                    brightness:brightness
+                         alpha:alpha];
 }
 
 @end
